+++ opentelemetry-api/src/opentelemetry/_logs/_internal/__init__.py
@@ -76,6 +76,7 @@ class LogRecord(ABC):
         body: AnyValue = None,
         attributes: Optional[_ExtendedAttributes] = None,
         event_name: Optional[str] = None,
+        exception: Optional[BaseException] = None,
     ) -> None: ...
 
     @overload
@@ -110,6 +111,7 @@ class LogRecord(ABC):
         body: AnyValue = None,
         attributes: Optional[_ExtendedAttributes] = None,
         event_name: Optional[str] = None,
+        exception: Optional[BaseException] = None,
     ) -> None:
         if not context:
             context = get_current()
@@ -127,6 +129,7 @@ class LogRecord(ABC):
         self.body = body
         self.attributes = attributes
         self.event_name = event_name
+        self.exception = exception
 
 
 class Logger(ABC):
@@ -157,12 +160,15 @@ class Logger(ABC):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     @overload
     def emit(
         self,
         record: LogRecord,
+        *,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     @abstractmethod
@@ -178,6 +184,7 @@ class Logger(ABC):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None:
         """Emits a :class:`LogRecord` representing a log to the processing pipeline."""
 
@@ -200,12 +207,15 @@ class NoOpLogger(Logger):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     @overload
     def emit(  # pylint:disable=arguments-differ
         self,
         record: LogRecord,
+        *,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     def emit(
@@ -220,6 +230,7 @@ class NoOpLogger(Logger):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None:
         pass
 
@@ -266,12 +277,15 @@ class ProxyLogger(Logger):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     @overload
     def emit(  # pylint:disable=arguments-differ
         self,
         record: LogRecord,
+        *,
+        exception: BaseException | None = None,
     ) -> None: ...
 
     def emit(
@@ -286,9 +300,10 @@ class ProxyLogger(Logger):
         body: AnyValue | None = None,
         attributes: _ExtendedAttributes | None = None,
         event_name: str | None = None,
+        exception: BaseException | None = None,
     ) -> None:
         if record:
-            self._logger.emit(record)
+            self._logger.emit(record, exception=exception)
         else:
             self._logger.emit(
                 timestamp=timestamp,
@@ -299,6 +314,7 @@ class ProxyLogger(Logger):
                 body=body,
                 attributes=attributes,
                 event_name=event_name,
+                exception=exception,
             )